refactor: remove comments and update types
