+++ app/components/OurImpactSection/types.ts
@@ -1,0 +1,4 @@
+export interface Impact {
+  result: string
+  description: string
+}
