+++ app/components/OurValuesSection/index.tsx
@@ -1,31 +1,11 @@
-// 'use client'
-
 import { Tables } from '@/app/protocols'
-// import { getContentfulData } from '@/app/utils/getContentfulData'
 import Image from 'next/image'
 
 import style from './style.module.css'
-
-export interface Values {
-  title: string
-  description: string
-  symbol: {
-    fields: {
-      file: {
-        url: string
-        details: {
-          image: {
-            width: number
-            height: number
-          }
-        }
-      }
-    }
-  }
-}
+import { Values } from './types'
 
 interface OurValuesProps {
-  values: Tables<Values>[] // Declara a propriedade values como um array de objetos Values
+  values: Tables<Values>[]
   tableName: string
 }
