+++ app/components/OurValuesSection/types.ts
@@ -1,0 +1,17 @@
+export interface Values {
+  title: string
+  description: string
+  symbol: {
+    fields: {
+      file: {
+        url: string
+        details: {
+          image: {
+            width: number
+            height: number
+          }
+        }
+      }
+    }
+  }
+}
